requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
